+++ udf/agent/py2/kapacitor/udf/agent.py
@@ -16,8 +16,8 @@ except ImportError:
 
 
 # Setup default in/out io
-defaultIn = sys.stdin.buffer
-defaultOut = sys.stdout.buffer
+defaultIn = sys.stdin
+defaultOut = sys.stdout
 
 import io
 import traceback
@@ -27,6 +27,23 @@ import struct
 
 import logging
 logger = logging.getLogger()
+
+# Check for python3
+# https://stackoverflow.com/a/38939320/703144
+if sys.version_info >= (3, 0):
+    defaultIn = sys.stdin.buffer
+    defaultOut = sys.stdout.buffer
+elif sys.version_info >= (2, 0):
+    logger.warning("[WARNING] DEPRECATED VERSION: Python2 version %d.%d.%d detected. "
+                   "Support for Python 2-based UDFs is deprecated as of Kapacitor 1.7.7 and "
+                   "will be removed in Kapacitor 1.8.0. Please update your UDFs to be "
+                   "Python 3-compatible before upgrading. This change is part of our effort to "
+                   "follow modern security best practices.",
+                   sys.version_info.major,
+                   sys.version_info.minor,
+                   sys.version_info.minor)
+else:
+    logger.error("[ERROR] Unsupported Python version %s detected", sys.version_info)
 
 # The Agent calls the appropriate methods on the Handler as requests are read off STDIN.
 #