+++ CHANGELOG.md
@@ -1,4 +1,14 @@
-CHANGELOG for Kapacitor master commit 71a5e1e036b81a2f4cd4de56fdf298090d0ee13a
+## v1.8.0 [unreleased]
+----------------------
+
+⚠️ Python based UDFs no longer automatically support Python2.  This change is part of our effort to follow modern security best practices. Note that as a courtesy to some users legacy Python2 compatible UDF agent code has been moved to the directory `/udf/agent/py2`. Further details can be found in the `README.md` in that directory.
+
+### Other
+
+1. [2845](https://github.com/influxdata/kapacitor/pull/2845): Python2 UDFs no longer supported and CI changes
+   1. Internal build image reverted to `cross-builder`
+   2. Reference GO version upgraded to 1.23.9
+   3. Python based UDFs no longer automatically support Python2. This change is part of our effort to follow modern security best practices. Note that as a courtesy to some users legacy Python2 compatible UDF agent code has been moved to the directory `/udf/agent/py2`. Further details can be found in the `README.md` in that directory.
 
 ## v1.7.7 [2025-05-27]
 ----------------------